major cleanup in rate distortion
